+++ subscribers/createRegion.js
@@ -2,7 +2,7 @@ module.exports = (kojo, logger) => {
 
     const {stair, mongo, config: {databases: {regions}}} = kojo.get();
 
-    stair.read('region.create', async (regionData) => {
+    stair.read('createRegion', async (regionData) => {
         const index = mongo.db(regions).collection('index');
         const {ops: [newRegion]} = await index.insertOne(regionData);
         logger.info('created:', newRegion);