+++ subscribers/produce.js
@@ -1,9 +1,9 @@
 module.exports = (kojo, logger) => {
 
     const stair = kojo.get('stair');
-    stair.read('produce', async (message, done) => {
+    stair.read('produce', async (payload, done) => {
         const produce = kojo.module('produce');
-        await produce.addToIds(message);
+        await produce.addToIds(payload);
         done();
     });
 };